Поправил тип данных в векторе. Добавил тестирование основного сценария.

+++ lw3/task1.v7/main.cpp
@@ -49,7 +49,7 @@ bool CompareNumbers(double i, double j)
 
 int main(int argc, char* argv[])
 {
-    vector<int> numbers;
+    vector<double> numbers;
     double number;
 
     while (cin >> number)
